Intranet: Correcció a esborrar horaris
git-svn-id: https://servidor.casadecultura.cat/public/ccg/trunk/intranet@311 6722b150-0e3f-45da-bd85-83856719fb75

+++ apps/intranet/modules/gestio/actions/actions.class.php
@@ -1316,8 +1316,10 @@ class gestioActions extends sfActions
 
     			$RH = $request->getParameter('horaris');    			
     			$OH = HorarisPeer::retrieveByPK($RH['HorarisID']);
-    			$this->getUser()->addLogAction($accio,'gActivitats',$OH);
-    			$OH->delete();    			    			
+    			if($OH instanceof Horaris):
+	    			$this->getUser()->addLogAction($accio,'gActivitats',$OH);
+	    			$OH->delete();    			    			
+	    		endif; 
 	   			$this->redirect('gestio/gActivitats?accio=HORARI');
 	   			
     		break;
